Add files via upload
This is a reup of the DHH v1.15 graphic control mod from Discord. DHH mods are over because "Connector JP" has quit the mod release scene and all his repos are down. Just plopping this here for the sake of preservation of a great lighting and effects controller mod for the AI-Girl/Ai-Syoujyo game. 

Must download all files and extract with 7zip, or anything capable of extracting 7z files. I chose 7zip because it is efficient and free. The original zip is inside and if it still has the password request, it is "zodgame" but I do not believe it does as I think I zipped it myself after extracting. 

None of this work is my own, nor do I claim any credit for it's creation. This is a free distribution, and may be redistributed freely, but should not be sold/resold in any form, except personal art creations derived from the use of the mod. 

All credit for this mod goes to "Connector_JP"